Fix tests in AuthActivityTest
