+++ app/src/androidTest/java/com/github/versus/EntryActivityTest.java
@@ -19,8 +19,16 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
+
+import static com.github.versus.utils.auth.EmulatorUserProvider.validMail;
+import static com.github.versus.utils.auth.EmulatorUserProvider.validPassword;
+import static org.junit.Assert.fail;
 
 import android.content.Intent;
 
@@ -103,5 +111,32 @@ public class EntryActivityTest {
 
         }
     }
+
+    /*
+    //@Test
+    public void withCredentialCached(){
+        // HR : Make sure the credentials are cached
+        Task<AuthResult> task = authenticator.signInWithMail(validMail(), validPassword());
+        // HR : Spin and wait for completion
+        while (!task.isComplete() && !task.isCanceled());
+        // HR : Verification part
+        if (task.isCanceled())
+            fail("Sign in was cancelled");
+        else if(!task.isSuccessful())
+            fail("Sign in failed");
+        else{
+            // HR : reload the activity and check that the main_activity is visible
+            scenario.getScenario().recreate();
+            onView(withId(R.id.main_activity_layout)).check(matches(isDisplayed()));
+        }
+
+    }
+
+    //@Test
+    public void withCredentialUncached(){
+        // HR : Check that the auth_fragment is visible
+        onView(withId(R.id.auth_fragment)).check(matches(isDisplayed()));
+    }
+    */
 
 }
